Added cdk nag and run unit test script for coverage.

+++ deployment/build-s3-dist.sh
@@ -69,6 +69,14 @@ npm uninstall -g aws-cdk && npm install -g aws-cdk
 cd $source_dir/constructs
 npm install
 cdk synth --output=$staging_dist_dir
+if [ $? -ne 0 ]
+then
+    echo "******************************************************************************"
+    echo "cdk-nag found errors"
+    echo "******************************************************************************"
+    exit 1
+fi
+
 cd $staging_dist_dir
 rm tree.json manifest.json cdk.out
